requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==4.0.0
Sphinx
django-extensions==1.6.6
Werkzeug==0.10.4
django-test-plus==1.0.12
factory_boy==2.5.2

# django-debug-toolbar that works with Django 1.5+
django-debug-toolbar==1.4.0

# improved REPL
ipdb==0.8.1

#getting rid of InsecurePlatformWarning
ndg-httpsclient==0.4.0
